1. Added private helper methods for Lehman-encoding a permutation as a hex string, or decoding a hex string into a permutation of integers.
2. Added "hex" attribute to the Genome class, along with initialization
in the constructor and a setter method, and a private parsing method
to convert the hex representation into a data structure that gives
each rotation for each job, taking into account date swaps.

The hex representation encodes a permutation of workers for each job,
and a permutation of dates representing "swaps" of duties by assigned
workers.

If one job requires more than one person, there's a separate set of
swaps for each one, but only one worker rotation. For the second and
subsequent slots for the same job, we assume that the same rotation
is used offset by some number.